feat(dtype): add Float16 (half-precision) support
Adds full Float16 type support across the engine, mirroring Float32/Float64:

- DataType::Float16 enum variant with arrow roundtrip, coercion rules, and Python bindings
- Float16Type/Float16Array type system (NumericNative, DaftNumericType, DaftFloatType)
- Literal::Float16(half::f16) with From/FromLiteral conversions and numpy interop
- All 9 type-dispatch macros, supertype coercion, and aggregation type rules
- Array ops: sort, sum, product, round, is_in, trigonometry, comparison, growable
- Series ops: abs, floor, log, pow, aggregations, downcast, serialization
- Hashing (NaN canonicalization), groupby, search sorted
- Float functions (is_nan, not_nan, is_inf) and numeric functions (ceil, round, sign, exp)
- Utf8 → Float16 casting via Float32 intermediate
- NdArray::F16 variant with numpy half feature for tensor support
- SQL type mapping, HTML rendering, Parquet/JSON compatibility
- Python: DataType.float16, is_float16(), np.float16/pa.float16 mappings
- Arrow roundtrip tests and series test coverage

+++ tests/series/__init__.py
@@ -4,4 +4,4 @@ import pyarrow as pa
 
 ARROW_INT_TYPES = [pa.int8(), pa.uint8(), pa.int16(), pa.uint16(), pa.int32(), pa.uint32(), pa.int64(), pa.uint64()]
 ARROW_STRING_TYPES = [pa.string(), pa.large_string()]
-ARROW_FLOAT_TYPES = [pa.float32(), pa.float64()]
+ARROW_FLOAT_TYPES = [pa.float16(), pa.float32(), pa.float64()]
